Personal Finance Calculator

+++ python_introduction/finance_calculator.py
@@ -1,7 +1,7 @@
-montly_income = int(input("Enter your monthly income: "))
-monthly_expenses = int(input("Enter your total monthly expenses: "))
+montly_income = float(input("Enter your monthly income: "))
+monthly_expenses = float(input("Enter your total monthly expenses: "))
 
-monthly_savings = montly_income - monthly_expenses
+monthly_savings = int(float(montly_income) - float(monthly_expenses))
 print(f"Your monthly savings are: ${monthly_savings}.")
 
 projected_savings = int((monthly_savings * 12) + (monthly_savings * 12 * 0.05))
